+++ docs/components/configuration-steps.tsx
@@ -1,14 +1,14 @@
-import { Accordion, Accordions } from "fumadocs-ui/components/accordion";
-import { CodeBlock, Pre } from "fumadocs-ui/components/codeblock";
-import { Heading } from "fumadocs-ui/components/heading";
-import { Step, Steps } from "fumadocs-ui/components/steps";
+import { Accordion, Accordions } from 'fumadocs-ui/components/accordion'
+import { DynamicCodeBlock } from 'fumadocs-ui/components/dynamic-codeblock'
+import { Heading } from 'fumadocs-ui/components/heading'
+import { Step, Steps } from 'fumadocs-ui/components/steps'
 
 export type ConfigurationSteps = {
-  readonly commands?: boolean;
-  readonly config?: string;
-  readonly pkg: string;
-  readonly providers?: string[];
-};
+  readonly commands?: boolean
+  readonly config?: string
+  readonly pkg: string
+  readonly providers?: string[]
+}
 
 export const ConfigurationSteps = ({
   pkg,
@@ -22,59 +22,53 @@ export const ConfigurationSteps = ({
         <Step>
           <Heading as="h4">Installs {pkg}</Heading>
           <p>
-            Installs the <code>{pkg}</code> package using the detected package
-            manager.
+            Installs the <code>{pkg}</code> package using the detected package manager.
           </p>
         </Step>
         {providers && (
           <Step>
             <Heading as="h4">Registers providers</Heading>
             <p>
-              Registers the following service providers inside the{" "}
-              <code>adonisrc.ts</code> file
+              Registers the following service providers inside the <code>adonisrc.ts</code> file
             </p>
-            <CodeBlock className="lang" title="adonisrc.ts">
-              <Pre>
-                {"{\n"}
-                {"  providers: [\n"}
-                {"  // ...other providers\n"}
-                {providers.map(
-                  (provider) =>
-                    `  () => import("${pkg}/providers/${provider}")\n`,
-                )}
-                {"}"}
-              </Pre>
-            </CodeBlock>
+            <DynamicCodeBlock
+              code={`{
+  providers: [
+    // ...other providers
+    ${providers.map((provider) => `() => import("${pkg}/providers/${provider}")`).join(',\n')}
+  ]
+}`}
+              lang="ts"
+            />
           </Step>
         )}
         {commands && (
           <Step>
             <Heading as="h4">Registers commands</Heading>
             <p>
-              Registers the following commands inside the{" "}
-              <code>adonisrc.ts</code> file
+              Registers the following commands inside the <code>adonisrc.ts</code> file
             </p>
-            <CodeBlock className="lang" title="adonisrc.ts">
-              <Pre>
-                {"{\n"}
-                {"  commands: [\n"}
-                {"  // ...other commands\n"}
-                {`  () => import("${pkg}/commands")\n`}
-                {"}"}
-              </Pre>
-            </CodeBlock>
+            <DynamicCodeBlock
+              code={`{
+  commands: [
+    // ...other commands
+    () => import("${pkg}/commands")
+  ]
+}`}
+              lang="ts"
+            />
           </Step>
         )}
         {config && (
           <Step>
             <Heading as="h4">Generates configuration</Heading>
             <p>
-              A configuration file <code>config/{config}.ts</code> is generated
-              containing the default configuration.
+              A configuration file <code>config/{config}.ts</code> is generated containing the
+              default configuration.
             </p>
           </Step>
         )}
       </Steps>
     </Accordion>
   </Accordions>
-);
+)