+++ docs/eslint.config.mjs
@@ -1,15 +1,15 @@
-import next from "@foadonis/eslint-config/next.app.js";
+import next from '@foadonis/eslint-config/next.app.js'
 
 export default [
   {
-    ignores: [".source/*"],
+    ignores: ['.source/*'],
   },
   ...next,
   {
     languageOptions: {
       parserOptions: {
-        project: "./tsconfig.lint.json",
+        project: './tsconfig.lint.json',
       },
     },
   },
-];
+]