+++ packages/lucid-parser/tests/fixtures/data/column_primitives.ts
@@ -1,0 +1,12 @@
+import { BaseModel, column } from '@adonisjs/lucid/orm'
+
+export default class ColumnPrimitives extends BaseModel {
+  @column()
+  declare $string: string
+
+  @column()
+  declare $number: number
+
+  @column()
+  declare $boolean: boolean
+}